reverted a return value to int.
git-svn-id: https://www.seasar.org/svn/sandbox/s2robot/trunk@15 a24fa471-b002-4554-a963-9a0eab8552e2

+++ s2robot/src/main/java/org/seasar/robot/service/DataService.java
@@ -28,7 +28,7 @@ public interface DataService {
 
     public abstract void store(AccessResult accessResult);
 
-    public abstract long getCount(String sessionId);
+    public abstract int getCount(String sessionId);
 
     public abstract void delete(String sessionId);
